requirements.txt
pandas
aiohttp
python-telegram-bot
python-dotenv
httpx
openai
beautifulsoup4
selenium
notebook
jupyterlab
notebook
seaborn